Extract app source watcher out of DaemonServer into a module singleton
The AppSourceWatcher was a DaemonServer instance field, with the change handler
(handleAppSourceChange) and the setEnsureAppSourceWatcher DI callback wired in
the server. Move all of it into the app-source-watcher module: a singleton with
startAppSourceWatcher()/stopAppSourceWatcher() called from the daemon startup
and shutdown sequences, and the change handler co-located with the watcher.

ensureAppSourceWatcher() now drives the singleton directly, so the
setEnsureAppSourceWatcher indirection is gone.
